Serve tab with catch data

Serve tab with catch data

Usage

catchTab(input, output, session, params, logs, catch = NULL, ...)

catchTabUI(...)

Arguments

input

Reactive holding the inputs

output

Reactive holding the outputs

session

Shiny session

params

Reactive value holding updated MizerParams object

logs

Environment holding the log of steady states.

catch

Data frame holding binned observed catch data. The data can be binned either into length bins or weight bins. In the former case the data frame should have columns length and dl holding the start of the size bins in cm and the width of the size bins in cm respectively. In the latter case the data frame should have columns weight and dw holding the start of the size bins in grams and the width of the size bins in grams. The data frame also needs to have the columns species (the name of the species), gear (the name of the gear) and catch (the number of individuals of a particular species caught by a particular gear in a size bin).

...

Unused