Get diffusion rate

The diffusion rate \(d_i(w)\) for species \(i\) and weight \(w\) has contributions from the encounter of fish prey and of resource. This is determined by summing over all prey species and the resource spectrum and then integrating over all prey sizes \(w_p\), weighted by predation kernel \(\phi(w,w_p)\): $$ d_i(w) = (1-f_i(w))(\alpha_i(1-\psi_i(w)))^2\gamma_i(w) \int \left( \theta_{ip} N_R(w_p) + \sum_{j} \theta_{ij} N_j(w_p) \right) \phi_i(w,w_p) w_p^2 \, dw_p. $$ Here \(N_j(w)\) is the abundance density of species \(j\) and \(N_R(w)\) is the abundance density of resource. The overall prefactor \(\gamma_i(w)\) determines the predation power of the predator. It could be interpreted as a search volume and is set with the setSearchVolume() function. The predation kernel \(\phi(w,w_p)\) is set with the setPredKernel() function. The species interaction matrix \(\theta_{ij}\) is set with setInteraction() and the resource interaction vector \(\theta_{ip}\) is taken from the interaction_resource column in params@species_params. \(f(w)\) is the feeding level calculated with getFeedingLevel(). \(\psi(w)\) is the proportion of the available energy that is invested in reproduction instead of growth, as stored in params@psi.

Usage

getDiffusion(
  params,
  n = initialN(params),
  n_pp = initialNResource(params),
  n_other = initialNOther(params),
  t = 0,
  order = 2,
  ...
)

Arguments

params

A MizerParams object

n

A matrix of species abundances (species x size).

n_pp

A vector of the resource abundance by size

n_other

A list of abundances for other dynamical components of the ecosystem

t

The time for which to do the calculation (Not used by standard mizer rate functions but useful for extensions with time-dependent parameters.)

order

By setting this to integer values above 2 one can get the coefficients of higher-order terms in the expansion of the jump-growth equation. The default is 2, which gives the diffusion rate.

...

Unused

Value

A named two dimensional array (predator species x predator size) with the diffusion rates.

Details

The function returns values also for sizes outside the size-range of the species. These values should not be used, as they are meaningless.