Get metabolic respiration for each species

For each species, returns the rate at which energy is lost to metabolic respiration by all individuals of that species. Note that this is different from the respiration rate \(R\) used by Ecopath models, see getRespiration().

Usage

getMetabolicRespiration(params)

Arguments

params

A MizerParams object

Value

A named vector of respiration for each species

Details

The respiration rate \(K_i\) is calculated as $$K_i = \int_{w_0}^{w_{max}} k_i(w) N_i(w) dw$$ where \(k_i(w)\) is the metabolic rate of an individual of species \(i\) and weight \(w\) (calculated with getMetabolicRate) and \(N_i(w)\) is the number density of species \(i\) at weight \(w\).

Used internally by matchConsumption() to help scale metabolic loss to match observed consumption.

See also

Examples

getMetabolicRespiration(NS_params)
#>        Sprat      Sandeel       N.pout      Herring          Dab      Whiting 
#> 8.404771e+11 1.317525e+13 2.364933e+12 3.664966e+12 2.721456e+10 4.166850e+11 
#>         Sole      Gurnard       Plaice      Haddock          Cod       Saithe 
#> 2.173629e+11 1.891428e+11 3.077511e+12 1.351965e+12 6.024398e+11 9.007126e+11