Get production rate as defined by Ecopath

Unlike getTotalProduction(), this function follows the Ecopath definition whereby the production rate in a mass-balanced model is equal to the rate at which biomass is removed by mortality. Thus the production rate includes the somatic production rate that is due to growth of individuals and it includes the rate at which offspring biomass is produced during reproduction. It does not include the full rate of gonad production because most of that is lost during the inefficient reproduction process. The Ecopath production rate is the sum of the somatic production rate obtained with getSomaticProduction and the rate of production of offspring biomass obtained with getOffspringProduction.

Usage

getProduction(params)

Arguments

params

A MizerParams object

Value

A named vector of production for each species

See also

Examples

getProduction(NS_params)
#>        Sprat      Sandeel       N.pout      Herring          Dab      Whiting 
#> 4.393148e+11 2.385652e+12 1.622204e+12 2.090316e+12 1.432059e+10 3.022551e+11 
#>         Sole      Gurnard       Plaice      Haddock          Cod       Saithe 
#> 1.089480e+11 1.373372e+11 2.586618e+12 8.681021e+11 3.557180e+11 4.836288e+11