Get the efficiency of reproduction

For each species, returns the fraction of the energy invested into reproduction that arrives in the size spectrum as offspring biomass. For a MizerParams object this is calculated from the initial state, for a MizerSim object it is returned for every saved time.

Usage

getReproductiveEfficiency(object)

Arguments

object

A MizerParams or MizerSim object

Value

If called with a MizerParams object, a named vector with the reproductive efficiency of each species in the model. If called with a MizerSim object, an ArrayTimeBySpecies object (time x species) with the reproductive efficiency at each saved time. The values are dimensionless fractions.

Details

Mature individuals invest energy into reproduction at the total rate \(E_{R.i}\) (grams per year), obtained with getGonadicProduction(). Mizer assumes that half of that investment is made by females, so \(E_{R.i} / 2\) is what is available for making eggs. Offspring biomass is produced at the rate \(R_{dd.i} w_{0.i}\), obtained with getOffspringProduction(). The reproductive efficiency is the ratio of the two: $$\frac{R_{dd.i}\, w_{0.i}}{E_{R.i} / 2} = \epsilon_i \, \frac{R_{dd.i}}{R_{di.i}},$$ where \(\epsilon_i\) is the erepro species parameter and \(R_{di.i}\) is the density-independent rate of egg production, see mizer::getRDI(). The second factor \(R_{dd.i}/R_{di.i}\) is the fraction of eggs that survives density dependence and equals 1 - reproduction_level, see mizer::setBevertonHolt().

The function evaluates the right-hand side of the above equation, because that is the form mizer itself uses when it projects reproduction. The two sides agree exactly, whatever mizer::second_order_w() is set to, because getGonadicProduction() discretises the reproduction integral with the same quadrature that mizer::getRDI() uses.

A value above 1 is energetically impossible: the model would be producing more offspring biomass than the adults invested. Use plotReproductiveEfficiency() to see how far each species is from that limit and how much of the loss is due to density dependence. Species that do not reproduce at all (\(R_{di.i} = 0\)) get NA.

See also

Examples

getReproductiveEfficiency(celtic_params)
#>   Blue whiting       Boarfish            Cod        Haddock           Hake 
#>   1.318272e-01   7.750503e-02   5.924691e-04   8.517803e-06   1.654638e-01 
#>        Herring Horse mackerel       Mackerel         Megrim       Monkfish 
#>   2.987550e-02   1.868785e-01   5.189377e-02   3.387677e-03   1.638693e-03 
#>         Plaice    Red gurnard           Sole        Whiting 
#>   1.237044e-01   2.665808e-02   2.396251e-02   4.809613e-03