Get rate of biomass loss due to reproduction

For each species, returns the difference between the rate of gonad production (calculated with getGonadicProduction) and the rate of offspring biomass production (calculated with getOffspringProduction).

Usage

getReproductiveLoss(params)

Arguments

params

A MizerParams object

Value

A named vector of reproduction biomass loss for each species

See also

Examples

getReproductiveLoss(NS_params)
#>        Sprat      Sandeel       N.pout      Herring          Dab      Whiting 
#> 1.572291e+11 6.332264e+12 2.621121e+11 2.737641e+11 5.280980e+09 8.389938e+10 
#>         Sole      Gurnard       Plaice      Haddock          Cod       Saithe 
#> 2.161179e+10 1.883767e+09 3.843950e+10 6.081030e+10 2.604508e+11 9.614549e+10