Get respiration rate as defined by Ecopath

Unlike getMetabolicRespiration(), this function follows the Ecopath definition of respiration as the rate at which assimilated food is NOT used for production. This is expressed by the first master equation of Ecopath: $$Q = P + R + U$$ where \(Q\) is the consumption rate, \(P\) is the production rate, \(R\) is the respiration rate, and \(U\) is the unassimilated part of the consumption rate, i.e., \(U = (1 - \alpha) Q\), where \(\alpha\) is the assimilation efficiency. Solving this for \(R\) we get: $$R = \alpha Q - P$$

Usage

getRespiration(params)

Arguments

params

A MizerParams object

Value

A named vector of respiration for each species

Details

In mizer, this respiration rate has two components: the metabolic respiration and the loss due to gonad production that does not result in offspring biomass. Thus the respiration rate is the sum of the metabolic respiration rate (calculated with getMetabolicRespiration()) and the rate of biomass loss due to reproduction (calculated with getReproductiveLoss()).

See also

Examples

getRespiration(NS_params)
#>        Sprat      Sandeel       N.pout      Herring          Dab      Whiting 
#> 9.977062e+11 1.950752e+13 2.627045e+12 3.938730e+12 3.249554e+10 5.005844e+11 
#>         Sole      Gurnard       Plaice      Haddock          Cod       Saithe 
#> 2.389746e+11 1.910265e+11 3.115950e+12 1.412775e+12 8.628906e+11 9.968581e+11