Get somatic production for each species

For each species, returns the rate at which somatic biomass is produced by all individuals of that species.

Usage

getSomaticProduction(params)

Arguments

params

A MizerParams object

Value

A named vector of somatic production for each species

Details

The somatic production rate \(P_{s.i}\) is calculated as $$P_{s.i} = \int_{w_0}^{w_{max}} g_i(w) N_i(w) dw$$ where \(g(w)\) is the somatic growth rate of an individual of species \(i\) and weight \(w\) (calculated with getEGrowth) and \(N_i(w)\) is the number density of species \(i\) at weight \(w\).

See also

Examples

getSomaticProduction(NS_params)
#>        Sprat      Sandeel       N.pout      Herring          Dab      Whiting 
#> 4.385836e+11 2.385242e+12 1.612457e+12 2.089215e+12 1.430943e+10 3.017142e+11 
#>         Sole      Gurnard       Plaice      Haddock          Cod       Saithe 
#> 1.089094e+11 1.366080e+11 2.553888e+12 8.663644e+11 3.557097e+11 4.835171e+11