Enable predator–prey interactions using an interaction matrix

This function should be called with a non-interacting model (where there is only external mortality and external encounter) and an interaction matrix. The external encounter and mortality rates will then be decreased by the amount of the predation encounter and mortality that is created by that interaction matrix.

Usage

makeInteracting(params, interaction)

Arguments

params

A MizerParams object

interaction

An interaction matrix where entry (i, j) gives the strength with which species i preys on j.

Value

The modified MizerParams object

Details

If predation would exceed total mortality or encounter for a species, a warning is issued and negative values are truncated to zero. This can cause the resulting model to diverge from the steady state.

Typically used internally by matchDiet()