Set initial abundances to single-species steady state abundances

This first calculates growth and death rates that arise from the current initial abundances. Then it uses these growth and death rates to determine the steady-state abundances of the selected species.

Usage

steadySingleSpecies(
  params,
  species = NULL,
  keep = c("egg", "biomass", "number")
)

Arguments

params

A MizerParams object

species

The species to be selected. Optional. By default all target species are selected. A vector of species names, or a numeric vector with the species indices, or a logical vector indicating for each species whether it is to be selected (TRUE) or not.

keep

A string determining which quantity is to be kept constant. The choices are "egg" which keeps the egg density constant, "biomass" which keeps the total biomass of the species constant and "number" which keeps the total number of individuals constant.

Value

A MizerParams object in which the initial abundances of the selected species are changed to their single-species steady state abundances.

Details

If a non-zero emigration rate is set for the species and is chosen so high that the species cannot sustain it, then an error is thrown.

If the species parameters d_over_g is set, then the diffusion rate is calculated from the growth rate at the smallest size and the d_over_g parameter. If d_over_g is not set, then the diffusion is set to zero.

The result of applying this function is of course not a multi-species steady state, because after changing the abundances of the selected species the growth and death rates will have changed.