Launch shiny gadget for tuning parameters

The function opens a shiny gadget, an interactive web page. This page has a side panel with controls for various model parameters and a main panel with tabs for various diagnostic plots.

Usage

tuneParams(
  params,
  controls = c("predation", "fishing", "reproduction", "other", "interaction",
    "resource"),
  tabs = c("Spectra", "Abundance", "Growth", "Repro", "Catch", "Diet", "Death",
    "Resource", "Rates", "Sim"),
  match = c("none"),
  preserve = c("erepro", "reproduction_level", "R_max"),
  return_app = FALSE,
  ...
)

Arguments

params

MizerParams object to tune. If missing, the gadget tries to recover information from log files left over from aborted previous runs.

controls

A character vector of names of input parameter control sections that should be displayed in the sidebar.

tabs

A character vector of names of the tabs that should be displayed in the main section.

match

A character vector. Determines which quantities should be matched to observations each time the "steady" button is pressed. Possible entries are "growth" (using mizer::matchGrowth()), "biomass" (using mizer::matchBiomasses()) and "yield" (using mizer::matchYields()).

preserve

Specifies whether the reproduction_level should be preserved or the maximum reproduction rate R_max or the reproductive efficiency erepro (Default). See mizer::setBevertonHolt() for an explanation of the reproduction_level.

return_app

Boolean. For testing purposes only. When set to TRUE will return a shinyApp object instead of running the gadget.

...

Other params needed by individual tabs.

Value

The tuned MizerParams object

Details

This gadget is meant for tuning a model to steady state. It is not meant for tuning the dynamics of the model. That should be done in a second step using functions like setRmax() or changeResource().

There is an "Instructions" button near the top left of the gadget that gives you a quick overview of the user interface.

After you click the "Return" button in the side panel, the function will return the parameter object in the state at that time, with Rmax set to Inf and erepro set to the value it had after the last run to steady state.

At any time the gadget allows the user to download the current params object as an .rds file via the "Download" button.