Match the model's production rate to empirical observations/estimates

This function sets the respiration rate according to the first master equation of Ecopath, which is: $$Q = P + R + U$$ where \(Q\) is the consumption rate, \(P\) is the somatic production rate, \(R\) is the respiration rate, and \(U\) is the unassimilated part of the consumption rate, i.e., \(U = (1 - \alpha) Q\), where \(\alpha\) is the assimilation efficiency. Solving this for \(R\) we get: $$R = Q \alpha - P$$

Usage

matchProduction(params, tol = 0.1, max_iter = 10)

matchProductionOnce(params, steady = TRUE)

isProductionMatched(params, tol = 0.1)

Arguments

params

A MizerParams object

tol

The relative tolerance for matching observed production. Default is 0.1.

max_iter

The maximum number of iterations. Default is 10.

steady

Logical. If TRUE (default), returns the model to steady state after parameter adjustment using steadySingleSpecies() and matchBiomasses().

Value

A MizerParams object with the adjusted metabolic and reproductive parameters to match production.

A MizerParams object with updated parameters and optionally a new steady state.

Details

Ecopath does not explicitly account for the loss due to gonad production. However most of the energy invested into gonad production is lost due to the inefficiency of the reproductive process. Therefore in mizer we need to include this loss in the respiration rate. Thus $$R = K + G - R_dd * w_0$$ where \(K\) is the metabolic rate, \(G\) is the gonadic production rate, \(R_dd\) is the rate of offspring production, and \(w_0\) is the weight of an offspring at the size it enters the model. Thus \(G - R_dd * w_0\) is the net loss due to gonad production.

The matchProduction function calls matchProductionOnce() repeatedly until the production matches within the tolerance specified by the tol parameter. The function will return a warning if the maximum number of iterations is reached without converging.

See also